Ahmed Belhadj - Bootstrap II
Used the flex utility to vertically center and move the navigation to the far right of it's container.
Used the border radius utility to round the corners on the images named adventure.jpg and fun.jpg;
Used the the display utility sticky-top on the the container-fluid div.
Used the spacer utility to give enough top-margin or bottom-margin to any element that looked like they were too close together.